add confirm dialog

+++ src/components/atoms/BasketContent.vue
@@ -4,7 +4,9 @@
       <p class="mb-6 mt-6 text-center text-3xl font-pacifico text-neutral-700">
         Twój koszyk pełen pyszności
       </p>
-      <div class="grid grid-cols-5 border-4 text-center text-lg md:text-xl lg:text-2xl font-semibold mt-4">
+      <div
+        class="grid grid-cols-5 border-4 text-center text-lg md:text-xl lg:text-2xl font-semibold mt-4"
+      >
         <div>Donut</div>
         <div>Cena</div>
         <div>Ilość</div>
@@ -16,36 +18,52 @@
         :donut="product"
         :key="data.id"
       >
-        <div class="font-semibold cursor-pointer border-2 border-white hover:border-slate-700 hover:rounded-2xl">
+        <div
+          class="font-semibold cursor-pointer border-2 border-white hover:border-slate-300 hover:rounded"
+        >
           <router-link :to="`/donut/${data.product.id}`">
-          {{ data.product.taste }}
-          <img :src="data.product.image" class="w-20 h-20 m-auto" />
-        </router-link>
+            {{ data.product.taste }}
+            <img :src="data.product.image" class="w-20 h-20 m-auto" />
+          </router-link>
         </div>
         <div class="m-auto">{{ data.product.price }}</div>
-        <input v-model="data.quantity" type="number" class="w-12 h-12 m-auto text-center border border-slate-400 rounded cursor-pointer focus:bg-slate-200"/>
-        <div class="m-auto">{{ (data.product.price * data.quantity).toFixed(2) }}</div>
-        <button @click="removeDonut(data.id)"
-        class="text-sm md:text-xl border rounded border-neutral-900 hover:bg-red-500 w-14 h-8 md:w-24 md:h-10 m-auto font-pacifico">Usuń</button>
+        <input
+          v-model="data.quantity"
+          type="number"
+          class="w-12 h-12 m-auto text-center border border-slate-400 rounded cursor-pointer focus:bg-slate-200"
+        />
+        <div class="m-auto">
+          {{ (data.product.price * data.quantity).toFixed(2) }}
+        </div>
+        <button
+          @click="removeDonut(data.id)"
+          class="text-sm md:text-xl border rounded border-neutral-900 hover:bg-red-500 w-14 h-8 md:w-24 md:h-10 m-auto font-pacifico"
+        >
+          Usuń
+        </button>
       </div>
       <div
         class="grid grid-cols-4 border-black border-2 text-center font-bold h-10"
       >
-        <div class="col-span-2 my-auto text-sm md:text-xl lg:text-2xl">Podsumowanie zakupów: </div>
+        <div class="col-span-2 my-auto text-sm md:text-xl lg:text-2xl">
+          Podsumowanie zakupów:
+        </div>
         <div class="my-auto text-sm md:text-xl lg:text-2xl">{{ result }}</div>
-        <div class="my-auto text-sm md:text-xl lg:text-2xl">Razem zł: {{ total.toFixed(2) }}</div>
+        <div class="my-auto text-sm md:text-xl lg:text-2xl">
+          Razem zł: {{ total.toFixed(2) }}
+        </div>
       </div>
     </div>
     <div class="flex justify-between">
       <router-link to="/DonutShop">
         <button
-          class="border border-neutral-900  bg-orange-200 hover:bg-orange-300 rounded p-3 font-pacifico text-2xl ml-4"
+          class="border border-neutral-900 bg-orange-400 hover:bg-orange-500 rounded p-3 font-pacifico text-2xl ml-4"
         >
           Kontynuuj zakupy
         </button>
       </router-link>
       <button
-        class="border border-neutral-900 bg-orange-200 hover:bg-orange-300 rounded p-3 font-pacifico text-2xl mr-4"
+        class="border border-neutral-900 bg-orange-400 hover:bg-orange-500 rounded p-3 font-pacifico text-2xl mr-4"
       >
         Przejdź do kasy
       </button>
@@ -59,12 +77,6 @@ import { useProductStore } from "@/stores/products";
 import { computed } from "vue";
 export default {
   name: "BasketContent",
-  props: {
-    donut: {
-      type: Object
-    }
-  },
-
   setup() {
     const store = useBasketStore();
     const { basket } = store;
@@ -81,8 +93,10 @@ export default {
     });
 
     const removeDonut = (id) => {
-      store.deleteProduct(id)
-    }
+      if (confirm("Czy na pewno chcesz usunąć tego donuta?")) {
+        store.deleteProduct(id);
+      }
+    };
 
     basket.map((el) => {
       el.product = donuts[el.id];
@@ -92,7 +106,7 @@ export default {
       donuts,
       result,
       total,
-      removeDonut
+      removeDonut,
     };
   },
 };
